+++ packages/cli/__tests__/zosfiles/__unit__/edit/Edit.handler.unit.test.ts
@@ -16,14 +16,13 @@ import { mockHandlerParameters } from "@zowe/cli-test-utils";
 import { EditDefinition } from "../../../../src/zosfiles/edit/Edit.definition";
 import EditHandler from "../../../../src/zosfiles/edit/Edit.handler";
 import { UNIT_TEST_PROFILES_ZOSMF, UNIT_TEST_ZOSMF_PROF_OPTS } from "../../../../../../__tests__/__src__/mocks/ZosmfProfileMock";
-import * as path from 'path';
-import * as os from 'os';
+import stripAnsi = require("strip-ansi");
 
 describe("Files Edit Group Handler", () => {
     describe("process method", () => {
         //Variable instantiation
         const dataSetName = "dataset";
-        const dataSetPath = path.join(process.cwd(), "packages\\cli\\src\\zosfiles\\edit\\Edit.handler.ts");
+        const dataSetPath = "/tmp/dataset.txt";
 
         const commandParameters: IHandlerParameters = mockHandlerParameters({
             arguments: UNIT_TEST_ZOSMF_PROF_OPTS,
@@ -52,7 +51,6 @@ describe("Files Edit Group Handler", () => {
         const guiAvailSpy = jest.spyOn(ProcessUtils, "isGuiAvailable");
         jest.spyOn(EditUtilities, "fileComparison").mockImplementation(jest.fn());
         jest.spyOn(EditUtilities, "makeEdits").mockImplementation(jest.fn());
-        jest.spyOn(os,"tmpdir").mockReturnValue("/tmp");
 
         guiAvailSpy.mockImplementation(jest.fn(() => {
             return GuiResult.GUI_AVAILABLE;
@@ -75,6 +73,9 @@ describe("Files Edit Group Handler", () => {
         commandParameters.arguments.dataSetName = dataSetName;
         const params = Object.assign({}, ...[commandParameters]);
         params.arguments = Object.assign({}, ...[commandParameters.arguments]);
+        params.response.console.log = jest.fn((logs) => {
+            expect(stripAnsi(logs.toString())).toMatchSnapshot();
+        }) as any;
 
         it("should make and upload edits successfully", async () => {
             jest.spyOn(EditUtilities, "makeEdits").mockImplementation(async () => {