+++ packages/zosfiles/__tests__/__unit__/methods/download/Download.unit.test.ts
@@ -21,6 +21,7 @@ import * as util from "util";
 import { IUSSListOptions, List } from "../../../../src/methods/list";
 import { CLIENT_PROPERTY } from "../../../../src/doc/types/ZosmfRestClientProperties";
 import { IDownloadDsmResult } from "../../../../src/methods/download/doc/IDownloadDsmResult";
+import { PassThrough } from "stream";
 
 describe("z/OS Files - Download", () => {
     const dsname = "USER.DATA.SET";
@@ -128,7 +129,7 @@ describe("z/OS Files - Download", () => {
             expect(caughtError).toBeUndefined();
             expect(response).toEqual({
                 success: true,
-                commandResponse: util.format(ZosFilesMessages.datasetDownloadedSuccessfully.message, destination),
+                commandResponse: util.format(ZosFilesMessages.datasetDownloadedWithDestination.message, destination),
                 apiResponse: {}
             });
 
@@ -165,7 +166,7 @@ describe("z/OS Files - Download", () => {
             expect(caughtError).toBeUndefined();
             expect(response).toEqual({
                 success: true,
-                commandResponse: util.format(ZosFilesMessages.datasetDownloadedSuccessfully.message, destination),
+                commandResponse: util.format(ZosFilesMessages.datasetDownloadedWithDestination.message, destination),
                 apiResponse: {}
             });
 
@@ -199,7 +200,7 @@ describe("z/OS Files - Download", () => {
             expect(caughtError).toBeUndefined();
             expect(response).toEqual({
                 success: true,
-                commandResponse: util.format(ZosFilesMessages.datasetDownloadedSuccessfully.message, destination),
+                commandResponse: util.format(ZosFilesMessages.datasetDownloadedWithDestination.message, destination),
                 apiResponse: {}
             });
 
@@ -235,7 +236,7 @@ describe("z/OS Files - Download", () => {
             expect(caughtError).toBeUndefined();
             expect(response).toEqual({
                 success: true,
-                commandResponse: util.format(ZosFilesMessages.datasetDownloadedSuccessfully.message, file),
+                commandResponse: util.format(ZosFilesMessages.datasetDownloadedWithDestination.message, file),
                 apiResponse: {}
             });
 
@@ -273,7 +274,7 @@ describe("z/OS Files - Download", () => {
             expect(caughtError).toBeUndefined();
             expect(response).toEqual({
                 success: true,
-                commandResponse: util.format(ZosFilesMessages.datasetDownloadedSuccessfully.message, file),
+                commandResponse: util.format(ZosFilesMessages.datasetDownloadedWithDestination.message, file),
                 apiResponse: {}
             });
 
@@ -310,7 +311,7 @@ describe("z/OS Files - Download", () => {
             expect(caughtError).toBeUndefined();
             expect(response).toEqual({
                 success: true,
-                commandResponse: util.format(ZosFilesMessages.datasetDownloadedSuccessfully.message, destination),
+                commandResponse: util.format(ZosFilesMessages.datasetDownloadedWithDestination.message, destination),
                 apiResponse: {}
             });
 
@@ -346,7 +347,7 @@ describe("z/OS Files - Download", () => {
             expect(caughtError).toBeUndefined();
             expect(response).toEqual({
                 success: true,
-                commandResponse: util.format(ZosFilesMessages.datasetDownloadedSuccessfully.message, file),
+                commandResponse: util.format(ZosFilesMessages.datasetDownloadedWithDestination.message, file),
                 apiResponse: {}
             });
 
@@ -383,7 +384,7 @@ describe("z/OS Files - Download", () => {
             expect(caughtError).toBeUndefined();
             expect(response).toEqual({
                 success: true,
-                commandResponse: util.format(ZosFilesMessages.datasetDownloadedSuccessfully.message, destination),
+                commandResponse: util.format(ZosFilesMessages.datasetDownloadedWithDestination.message, destination),
                 apiResponse: {}
             });
 
@@ -417,7 +418,7 @@ describe("z/OS Files - Download", () => {
             expect(caughtError).toBeUndefined();
             expect(response).toEqual({
                 success: true,
-                commandResponse: util.format(ZosFilesMessages.datasetDownloadedSuccessfully.message, file),
+                commandResponse: util.format(ZosFilesMessages.datasetDownloadedWithDestination.message, file),
                 apiResponse: {}
             });
 
@@ -454,7 +455,7 @@ describe("z/OS Files - Download", () => {
             expect(caughtError).toBeUndefined();
             expect(response).toEqual({
                 success: true,
-                commandResponse: util.format(ZosFilesMessages.datasetDownloadedSuccessfully.message, destination),
+                commandResponse: util.format(ZosFilesMessages.datasetDownloadedWithDestination.message, destination),
                 apiResponse: {}
             });
 
@@ -491,7 +492,7 @@ describe("z/OS Files - Download", () => {
             expect(caughtError).toBeUndefined();
             expect(response).toEqual({
                 success: true,
-                commandResponse: util.format(ZosFilesMessages.datasetDownloadedSuccessfully.message, destination),
+                commandResponse: util.format(ZosFilesMessages.datasetDownloadedWithDestination.message, destination),
                 apiResponse: {etag: etagValue}
             });
 
@@ -540,6 +541,37 @@ describe("z/OS Files - Download", () => {
                 normalizeResponseNewLines: true,
                 task: undefined /*no progress task*/});
 
+        });
+
+        it("should download a data set to a stream", async () => {
+            let response;
+            let caughtError;
+            const responseStream = new PassThrough();
+
+            try {
+                response = await Download.dataSet(dummySession, dsname, { stream: responseStream });
+            } catch (e) {
+                caughtError = e;
+            }
+
+            const endpoint = posix.join(ZosFilesConstants.RESOURCE, ZosFilesConstants.RES_DS_FILES, dsname);
+
+            expect(caughtError).toBeUndefined();
+            expect(response).toEqual({
+                success: true,
+                commandResponse: ZosFilesMessages.datasetDownloadedSuccessfully.message,
+                apiResponse: {}
+            });
+
+            expect(zosmfGetFullSpy).toHaveBeenCalledTimes(1);
+            expect(zosmfGetFullSpy).toHaveBeenCalledWith(dummySession, {resource: endpoint,
+                reqHeaders: [ZosmfHeaders.ACCEPT_ENCODING, ZosmfHeaders.TEXT_PLAIN],
+                responseStream,
+                normalizeResponseNewLines: true,
+                task: undefined /*no progress task*/});
+
+            expect(ioCreateDirSpy).not.toHaveBeenCalled();
+            expect(ioWriteStreamSpy).not.toHaveBeenCalled();
         });
     });
 
@@ -642,7 +674,7 @@ describe("z/OS Files - Download", () => {
             expect(caughtError).toBeUndefined();
             expect(response).toEqual({
                 success: true,
-                commandResponse: util.format(ZosFilesMessages.datasetDownloadedSuccessfully.message, dsFolder),
+                commandResponse: util.format(ZosFilesMessages.datasetDownloadedWithDestination.message, dsFolder),
                 apiResponse: listApiResponse
             });
 
@@ -675,7 +707,7 @@ describe("z/OS Files - Download", () => {
             expect(caughtError).toBeUndefined();
             expect(response).toEqual({
                 success: true,
-                commandResponse: util.format(ZosFilesMessages.datasetDownloadedSuccessfully.message, directory),
+                commandResponse: util.format(ZosFilesMessages.datasetDownloadedWithDestination.message, directory),
                 apiResponse: listApiResponse
             });
 
@@ -710,7 +742,7 @@ describe("z/OS Files - Download", () => {
             expect(caughtError).toBeUndefined();
             expect(response).toEqual({
                 success: true,
-                commandResponse: util.format(ZosFilesMessages.datasetDownloadedSuccessfully.message, directory),
+                commandResponse: util.format(ZosFilesMessages.datasetDownloadedWithDestination.message, directory),
                 apiResponse: listApiResponse
             });
 
@@ -746,7 +778,7 @@ describe("z/OS Files - Download", () => {
             expect(caughtError).toBeUndefined();
             expect(response).toEqual({
                 success: true,
-                commandResponse: util.format(ZosFilesMessages.datasetDownloadedSuccessfully.message, directory),
+                commandResponse: util.format(ZosFilesMessages.datasetDownloadedWithDestination.message, directory),
                 apiResponse: listApiResponse
             });
 
@@ -784,7 +816,7 @@ describe("z/OS Files - Download", () => {
             expect(caughtError).toBeUndefined();
             expect(response).toEqual({
                 success: true,
-                commandResponse: util.format(ZosFilesMessages.datasetDownloadedSuccessfully.message, directory),
+                commandResponse: util.format(ZosFilesMessages.datasetDownloadedWithDestination.message, directory),
                 apiResponse: listApiResponse
             });
 
@@ -821,7 +853,7 @@ describe("z/OS Files - Download", () => {
             expect(caughtError).toBeUndefined();
             expect(response).toEqual({
                 success: true,
-                commandResponse: util.format(ZosFilesMessages.datasetDownloadedSuccessfully.message, directory),
+                commandResponse: util.format(ZosFilesMessages.datasetDownloadedWithDestination.message, directory),
                 apiResponse: listApiResponse
             });
 
@@ -856,7 +888,7 @@ describe("z/OS Files - Download", () => {
             expect(caughtError).toBeUndefined();
             expect(response).toEqual({
                 success: true,
-                commandResponse: util.format(ZosFilesMessages.datasetDownloadedSuccessfully.message, directory),
+                commandResponse: util.format(ZosFilesMessages.datasetDownloadedWithDestination.message, directory),
                 apiResponse: listApiResponse
             });
 
@@ -886,7 +918,7 @@ describe("z/OS Files - Download", () => {
             expect(caughtError).toBeUndefined();
             expect(response).toEqual({
                 success: true,
-                commandResponse: util.format(ZosFilesMessages.datasetDownloadedSuccessfully.message, dsFolder.toUpperCase()),
+                commandResponse: util.format(ZosFilesMessages.datasetDownloadedWithDestination.message, dsFolder.toUpperCase()),
                 apiResponse: listApiResponse
             });
 
@@ -1514,7 +1546,7 @@ describe("z/OS Files - Download", () => {
                             { member: "TESTDS" }
                         ]
                     },
-                    commandResponse: util.format(ZosFilesMessages.datasetDownloadedSuccessfully.message, "./")
+                    commandResponse: util.format(ZosFilesMessages.datasetDownloadedWithDestination.message, "./")
                 };
             });
 
@@ -1545,7 +1577,7 @@ describe("z/OS Files - Download", () => {
                 }, {}),
                 apiResponse: [{
                     ...dataSetPO,
-                    status: util.format(ZosFilesMessages.datasetDownloadedSuccessfully.message, "./") + "\nMembers:  TESTDS;"
+                    status: util.format(ZosFilesMessages.datasetDownloadedWithDestination.message, "./") + "\nMembers:  TESTDS;"
                 }]
             });
         });
@@ -1608,7 +1640,7 @@ describe("z/OS Files - Download", () => {
                             { member: "TESTDS" }
                         ]
                     },
-                    commandResponse: util.format(ZosFilesMessages.datasetDownloadedSuccessfully.message, directory)
+                    commandResponse: util.format(ZosFilesMessages.datasetDownloadedWithDestination.message, directory)
                 };
             });
 
@@ -1639,7 +1671,7 @@ describe("z/OS Files - Download", () => {
                 }, {directory}),
                 apiResponse: [{
                     ...dataSetPO,
-                    status: util.format(ZosFilesMessages.datasetDownloadedSuccessfully.message, directory) + "\nMembers:  TESTDS;"
+                    status: util.format(ZosFilesMessages.datasetDownloadedWithDestination.message, directory) + "\nMembers:  TESTDS;"
                 }]
             });
         });
@@ -1748,7 +1780,7 @@ describe("z/OS Files - Download", () => {
             expect(caughtError).toBeUndefined();
             expect(response).toEqual({
                 success: true,
-                commandResponse: util.format(ZosFilesMessages.ussFileDownloadedSuccessfully.message, destination),
+                commandResponse: util.format(ZosFilesMessages.ussFileDownloadedWithDestination.message, destination),
                 apiResponse: {}
             });
 
@@ -1782,7 +1814,7 @@ describe("z/OS Files - Download", () => {
             expect(caughtError).toBeUndefined();
             expect(response).toEqual({
                 success: true,
-                commandResponse: util.format(ZosFilesMessages.ussFileDownloadedSuccessfully.message, destination),
+                commandResponse: util.format(ZosFilesMessages.ussFileDownloadedWithDestination.message, destination),
                 apiResponse: {}
             });
 
@@ -1820,7 +1852,7 @@ describe("z/OS Files - Download", () => {
             expect(caughtError).toBeUndefined();
             expect(response).toEqual({
                 success: true,
-                commandResponse: util.format(ZosFilesMessages.ussFileDownloadedSuccessfully.message, destination),
+                commandResponse: util.format(ZosFilesMessages.ussFileDownloadedWithDestination.message, destination),
                 apiResponse: {}
             });
 
@@ -1858,7 +1890,7 @@ describe("z/OS Files - Download", () => {
             expect(caughtError).toBeUndefined();
             expect(response).toEqual({
                 success: true,
-                commandResponse: util.format(ZosFilesMessages.ussFileDownloadedSuccessfully.message, destination),
+                commandResponse: util.format(ZosFilesMessages.ussFileDownloadedWithDestination.message, destination),
                 apiResponse: {}
             });
 
@@ -1899,7 +1931,7 @@ describe("z/OS Files - Download", () => {
             expect(caughtError).toBeUndefined();
             expect(response).toEqual({
                 success: true,
-                commandResponse: util.format(ZosFilesMessages.ussFileDownloadedSuccessfully.message, destination),
+                commandResponse: util.format(ZosFilesMessages.ussFileDownloadedWithDestination.message, destination),
                 apiResponse: {}
             });
 
@@ -1935,7 +1967,7 @@ describe("z/OS Files - Download", () => {
             expect(caughtError).toBeUndefined();
             expect(response).toEqual({
                 success: true,
-                commandResponse: util.format(ZosFilesMessages.ussFileDownloadedSuccessfully.message, destination),
+                commandResponse: util.format(ZosFilesMessages.ussFileDownloadedWithDestination.message, destination),
                 apiResponse: {}
             });
 
@@ -1968,7 +2000,7 @@ describe("z/OS Files - Download", () => {
             expect(caughtError).toBeUndefined();
             expect(response).toEqual({
                 success: true,
-                commandResponse: util.format(ZosFilesMessages.ussFileDownloadedSuccessfully.message, file),
+                commandResponse: util.format(ZosFilesMessages.ussFileDownloadedWithDestination.message, file),
                 apiResponse: {}
             });
 
@@ -2004,7 +2036,7 @@ describe("z/OS Files - Download", () => {
             expect(caughtError).toBeUndefined();
             expect(response).toEqual({
                 success: true,
-                commandResponse: util.format(ZosFilesMessages.ussFileDownloadedSuccessfully.message, destination),
+                commandResponse: util.format(ZosFilesMessages.ussFileDownloadedWithDestination.message, destination),
                 apiResponse: {etag: etagValue}
             });
 
@@ -2020,6 +2052,37 @@ describe("z/OS Files - Download", () => {
 
             expect(ioWriteStreamSpy).toHaveBeenCalledTimes(1);
             expect(ioWriteStreamSpy).toHaveBeenCalledWith(destination);
+        });
+
+        it("should download uss file to a stream", async () => {
+            let response;
+            let caughtError;
+            const responseStream = new PassThrough();
+
+            try {
+                response = await Download.ussFile(dummySession, ussname, { stream: responseStream });
+            } catch (e) {
+                caughtError = e;
+            }
+
+            const endpoint = posix.join(ZosFilesConstants.RESOURCE, ZosFilesConstants.RES_USS_FILES, encodeURIComponent(ussname.substr(1)));
+
+            expect(caughtError).toBeUndefined();
+            expect(response).toEqual({
+                success: true,
+                commandResponse: ZosFilesMessages.ussFileDownloadedSuccessfully.message,
+                apiResponse: {}
+            });
+
+            expect(zosmfGetFullSpy).toHaveBeenCalledTimes(1);
+            expect(zosmfGetFullSpy).toHaveBeenCalledWith(dummySession, {resource: endpoint,
+                reqHeaders: [ZosmfHeaders.ACCEPT_ENCODING, ZosmfHeaders.TEXT_PLAIN],
+                responseStream,
+                normalizeResponseNewLines: true,
+                task: undefined /* no progress task */});
+
+            expect(ioCreateDirSpy).not.toHaveBeenCalled();
+            expect(ioWriteStreamSpy).not.toHaveBeenCalled();
         });
     });
 
@@ -2398,6 +2461,8 @@ describe("z/OS Files - Download", () => {
             });
             expect(Download.ussFile).toHaveBeenCalledTimes(2);
         });
+
+        // TODO should download a uss file to a stream
     });
 
     describe("buildDownloadDsmResponse", () => {