+++ packages/zosfiles/CHANGELOG.md
@@ -3,7 +3,9 @@
 All notable changes to the Zowe z/OS files SDK package will be documented in this file.
 
 ## Recent Changes
-- Enhancement: Created zos-files edit commands to edit a dataset or uss file locally [PR #1672](https://github.com/zowe/zowe-cli/pull/1672)
+
+- Enhancement: Added streaming capabilities to the `Download.dataSet` and `Download.ussFile` methods.
+- BugFix: Fixed `Get.USSFile` API not respecting USS file tags.
 
 ## `7.16.4`